Significant imporvements to the C# implementation for server side.
Fixes the issue with nested properties which will now decode from Property.OtherProperty notation correctly.

Added LoadCriteria.FromExpressions that allows type safe filtering. See the TypeSafeLoad test for how to use it with type safe ordering as well.

This allows you to build a LoadCriteria in C# clients or elsewhere that will always be type safe.